Updated Route Card Editing

+++ src/Client/Pages/Progress/MembersDialog.razor.cs
@@ -53,7 +53,7 @@ namespace EDO_FOMS.Client.Pages.Progress
 
         private async Task<IEnumerable<ContactResponse>> SearchMembersAsync(string search)
         {
-            var response = await DocManager.GetAgreementMembersAsync(_agreement.EmplOrgId, search);
+            var response = await DocManager.GetAgreementMembersAsync((int)_agreement.EmplOrgId, search);
             return (response.Succeeded) ? response.Data : new();
         }
